requirements.txt
django-ckeditor
django-suit
django>2
djangorestframework
wheel
psycopg2